projects/memagents/pyproject.toml
[project]
name = "memagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.13",
    "ipykernel>=6.29.5",
    "llama-index-core>=0.12.43",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-llms-openai>=0.4.7",
    "llama-index-memory-mem0>=0.3.1",
    "llama-index-vector-stores-chroma>=0.4.2",
    "llama-index-vector-stores-qdrant>=0.6.1",
    "openai>=1.93.0",
    "protobuf>=3.20.0,<3.21.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.23.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.3",
]

[tool.uv.workspace]
members = [
]
